refactor(Participant): Participant가 이름을 갖도록 수정
