+++ src/main/java/blackjack/domain/participant/Dealer.java
@@ -1,17 +1,19 @@
 package blackjack.domain.participant;
 
 import blackjack.domain.card.Card;
+import blackjack.domain.card.Hand;
 import blackjack.domain.card.Score;
 import java.util.List;
 
 public class Dealer implements Decidable {
 
     private static final int DEALER_MIN_RECEIVE_CARD = 16;
+    private static final String DEALER_NAME = "딜러";
 
     private final Participant participant;
 
-    public Dealer(Participant participant) {
-        this.participant = participant ;
+    public Dealer() {
+        this.participant = new Participant(Hand.generateEmptyCards(), DEALER_NAME) ;
     }
 
     public void receiveCard(Card card) {
@@ -43,5 +45,9 @@ public class Dealer implements Decidable {
 
     public List<Card> getCards() {
         return participant.getHand();
+    }
+
+    public String getName() {
+        return participant.getName();
     }
 }
