+++ src/main/java/blackjack/domain/participant/Participant.java
@@ -7,10 +7,24 @@ import java.util.List;
 
 public class Participant {
 
-    private Hand hand;
+    private static final int MIN_NAME_LENGTH = 1;
+    private static final int MAX_NAME_LENGTH = 5;
 
-    public Participant(final Hand hand) {
+    private Hand hand;
+    private final String name;
+
+    public Participant(final Hand hand, final String name) {
+        validateNameLength(name);
         this.hand = hand;
+        this.name = name;
+    }
+
+    private void validateNameLength(final String name) {
+        if (name.length() < MIN_NAME_LENGTH || name.length() > MAX_NAME_LENGTH) {
+            throw new IllegalArgumentException(
+                    "[ERROR] 이름 길이는 최소 " + MIN_NAME_LENGTH + "글자에서 최대 "
+                            + MAX_NAME_LENGTH + "글자 입니다.");
+        }
     }
 
     public void receiveCard(Card card) {
@@ -31,5 +45,9 @@ public class Participant {
 
     public List<Card> getHand() {
         return hand.getHand();
+    }
+
+    public String getName() {
+        return name;
     }
 }
