+++ src/main/java/blackjack/domain/participant/Player.java
@@ -2,32 +2,20 @@ package blackjack.domain.participant;
 
 import blackjack.domain.Money;
 import blackjack.domain.card.Card;
+import blackjack.domain.card.Hand;
 import blackjack.domain.card.Score;
 import java.util.List;
 
 public class Player implements Decidable {
 
-    private static final int MIN_NAME_LENGTH = 1;
-    private static final int MAX_NAME_LENGTH = 5;
     private static final int PLAYER_MAX_RECEIVE_CARD = 21;
 
-    private final String name;
     private final Participant participant;
     private Money bettingMoney;
 
-    public Player(Participant participant, String name) {
-        validateNameLength(name);
-        this.name = name;
-        this.participant = participant;
+    public Player(String name) {
+        this.participant = new Participant(Hand.generateEmptyCards(), name);
         this.bettingMoney = Money.init();
-    }
-
-    private void validateNameLength(final String name) {
-        if (name.length() < MIN_NAME_LENGTH || name.length() > MAX_NAME_LENGTH) {
-            throw new IllegalArgumentException(
-                    "[ERROR] 이름 길이는 최소 " + MIN_NAME_LENGTH + "글자에서 최대 "
-                            + MAX_NAME_LENGTH + "글자 입니다.");
-        }
     }
 
     public void receiveCard(Card card) {
@@ -66,7 +54,7 @@ public class Player implements Decidable {
     }
 
     public String getName() {
-        return name;
+        return participant.getName();
     }
 
     public Money getBettingMoney() {
