+++ src/test/java/blackjack/domain/result/ProfitTest.java
@@ -26,7 +26,7 @@ class ProfitTest {
     @BeforeEach
     void setUp() {
         Players players = Players.from(List.of("a", "b"));
-        Dealer dealer = new Dealer(new Participant(Hand.generateEmptyCards()));
+        Dealer dealer = new Dealer();
         participants = new Participants(dealer, players);
         profit = new Profit(participants);
         bettingMoney();
